请求参数调整
