--- plugins/qrcode-plugin/src/main/java/com/github/hui/quick/plugin/qrcode/helper/v3/entity/QrResource.java
@@ -1,15 +1,0 @@
-package com.github.hui.quick.plugin.qrcode.helper.v3.entity;
-
-import java.awt.image.BufferedImage;
-
-/**
- * @author YiHui
- * @date 2022/7/20
- */
-public class QrResource {
-    private BufferedImage img;
-
-    private String text;
-
-    private String svg;
-}
