+++ plugins/qrcode-plugin/src/main/java/com/github/hui/quick/plugin/qrcode/v3/constants/TxtMode.java
@@ -1,0 +1,12 @@
+package com.github.hui.quick.plugin.qrcode.v3.constants;
+
+public enum TxtMode {
+    /***
+     * 文字二维码，随机模式
+     */
+    RANDOM,
+    /**
+     * 文字二维码，顺序模式
+     */
+    ORDER;
+}
