+++ plugins/qrcode-plugin/src/main/java/com/github/hui/quick/plugin/qrcode/v3/entity/render/RenderDot.java
@@ -1,6 +1,6 @@
-package com.github.hui.quick.plugin.qrcode.helper.v3.entity.render;
+package com.github.hui.quick.plugin.qrcode.v3.entity.render;
 
-import com.github.hui.quick.plugin.qrcode.helper.v3.entity.QrResource;
+import com.github.hui.quick.plugin.qrcode.v3.entity.QrResource;
 
 /**
  * 二维矩阵渲染
